Refactor exception to checked exception.
A ConflictException is a possible outcome when querying the DB and not a programming error. This is how we communicate to the caller that some input might result in an exception. We should make this clear in the signature of the functions and not hide it behind an unchecked exception.